Stop autodownload from running forever on map downloading errors.
The error status from low level of map download was not passed on.
Therefore update emits (to refrsh display) are generated even when
downloading got errors and no map tile needs to be redisplayed.

This caused serious problems when "Autodownload" is turned on.
The update emits also trigger autodownload to start new download thread
which in turn generates update emit for failed downloading of the
same tiles. And on and on.

This commit allow the download error status to be passed on and limit
update emits to successful download only.

Only google module was changed to use the new method. Other modules
will follow in the next commits.

Signed-off-by: Quy Tonthat <qtonthat@gmail.com>


git-svn-id: https://svn.sourceforge.net/svnroot/viking/trunk@236 02d69b8e-d120-0410-893f-fe71a3b6e39e