Google Directions now remembers previous query (from/to entries).
The from/to dialog will now display the previous query so that users can edit
or finetune the search.
(It was frustating mistyping a query just to get blank entriess back for
mistyping again, and again).
Note that this only works for one lifetime. Viking will forget the queries
once restarted.

Signed-off-by: Quy Tonthat <qtonthat@gmail.com>


git-svn-id: https://svn.sourceforge.net/svnroot/viking/trunk@252 02d69b8e-d120-0410-893f-fe71a3b6e39e