Fix intermittent problem of map redraw.
The type mismatching (gint vs guint16) in the cache removing function can
cause coords truncated which leads to pixmaps not removed from the cache,
hence they are not updated on screen.

This only happens for coords with values larger than what a 16-bit number
can hold (> 65535). This is why the problem seems intermittent.

Signed-off-by: Quy Tonthat <qtonthat@gmail.com>


git-svn-id: https://svn.sourceforge.net/svnroot/viking/trunk@262 02d69b8e-d120-0410-893f-fe71a3b6e39e