Add full support for GPX routes.
GPX Routes can be downloaded/uploaded from GPS devices.
When GPX files are imported/exported the route property is maintained for such tracks.
Allow converting of a track to/from a route via the layers track menu.
For any routes, draw in a different colour (the default is red) from other tracks. This colour is controllable in the Track Layer properties.

Internally, a route is implemented as a simple boolean flag as part of the track's property.
NB the gpx tags update in gpx.c also fixes the use of route names. (previously not considered).